Moved models package to data.models
